+++ src/main/java/com/axperty/stackedblocksfarmersdelight/registry/CreativeTabRegistry.java
@@ -22,6 +22,11 @@ public class CreativeTabRegistry {
                         entries.add(BlockRegistry.STACKED_COBBLESTONE_BLOCKS);
                         entries.add(BlockRegistry.STACKED_NETHERRACK_BLOCKS);
                         entries.add(BlockRegistry.STACKED_BRICKS);
+                        entries.add(BlockRegistry.STACKED_RESIN_BLOCKS);
+                        entries.add(BlockRegistry.STACKED_RESIN_BRICKS);
+                        entries.add(BlockRegistry.STACKED_MELONS);
+                        entries.add(BlockRegistry.STACKED_PUMPKINS);
+
                         entries.add(BlockRegistry.STACKED_RAW_IRON_BLOCKS);
                         entries.add(BlockRegistry.STACKED_RAW_GOLD_BLOCKS);
                         entries.add(BlockRegistry.STACKED_RAW_COPPER_BLOCKS);
@@ -35,9 +40,6 @@ public class CreativeTabRegistry {
                         entries.add(BlockRegistry.STACKED_DIAMOND_BLOCKS);
                         entries.add(BlockRegistry.STACKED_NETHERITE_BLOCKS);
                         entries.add(BlockRegistry.STACKED_QUARTZ_BLOCKS);
-                        entries.add(BlockRegistry.STACKED_MELONS);
-                        entries.add(BlockRegistry.STACKED_PUMPKINS);
-
 
                         entries.add(BlockRegistry.STACKED_OAK_LOGS);
                         entries.add(BlockRegistry.STACKED_STRIPPED_OAK_LOGS);
@@ -71,6 +73,10 @@ public class CreativeTabRegistry {
                         entries.add(BlockRegistry.STACKED_STRIPPED_CHERRY_LOGS);
                         entries.add(BlockRegistry.STACKED_CHERRY_PLANKS);
 
+                        entries.add(BlockRegistry.STACKED_PALE_OAK_LOGS);
+                        entries.add(BlockRegistry.STACKED_STRIPPED_PALE_OAK_LOGS);
+                        entries.add(BlockRegistry.STACKED_PALE_OAK_PLANKS);
+
                         entries.add(BlockRegistry.STACKED_BAMBOO_BLOCKS);
                         entries.add(BlockRegistry.STACKED_STRIPPED_BAMBOO_BLOCKS);
                         entries.add(BlockRegistry.STACKED_BAMBOO_PLANKS);
@@ -85,7 +91,5 @@ public class CreativeTabRegistry {
                     })
                     .build());
 
-    public static void registerItemGroups() {
-        StackedBlocksFarmersDelight.LOGGER.info("Registering item groups for " + StackedBlocksFarmersDelight.MODID);
-    }
+    public static void registerItemGroups() {}
 }